Fixed diff question MAS problem.

+++ lessons/mas_generator.py
@@ -8,7 +8,7 @@ variable_regex = r"\$\[\w+,\d+,\d+\]"
 def rand(seed, type, lower, upper): #Generates random values from a seed
     random.seed(seed)
     if type == "INT":
-        return float(random.randint(lower, upper))
+        return int(random.randint(lower, upper))
     elif type == "DEC":
         return float(lower + (random.random()*(upper-lower)))
     else:
